Added static analysis tool for detecting TA global and static variables using LLVM
The tool automatically detects when compiling the optee_example
and only allows the TA to be compiled if the detection is passed
(i.e. there are no custom global or static variables in the TA).
Since hpe_victim is an example of our hpe exploit for global variables,
we allow hpe_victim to bypass the static analysis tool's checks.